Trait unsafe_io::os::posish::AsRawReadWriteFd

pub trait AsRawReadWriteFd {
    fn as_raw_read_fd(&self) -> RawFd;

    fn as_raw_write_fd(&self) -> RawFd;
}

Like AsRawFd, but for types which may have one or two file descriptors, for reading and writing.

For types that only have one, both functions return the same value.

Required methods

fn as_raw_read_fd(&self) -> RawFd

Extracts the raw file descriptor for reading.

Like AsRawFd::as_raw_fd, but returns the reading file descriptor.

fn as_raw_write_fd(&self) -> RawFd

Extracts the raw file descriptor for writing.

Like AsRawFd::as_raw_fd, but returns the writing file descriptor.

Loading content...