Trait unsafe_io::os::posish::AsReadWriteFd

pub trait AsReadWriteFd<'f> {
    fn as_read_fd(self) -> BorrowedFd<'f>;

    fn as_write_fd(self) -> BorrowedFd<'f>;
}

Like AsFd, but for types which may have one or two file descriptors, for reading and writing.

For types that only have one, both functions return the same value.

Required methods

fn as_read_fd(self) -> BorrowedFd<'f>

Extracts the file descriptor for reading.

Like AsFd::as_fd, but returns the reading file descriptor.

fn as_write_fd(self) -> BorrowedFd<'f>

Extracts the file descriptor for writing.

Like AsFd::as_fd, but returns the writing file descriptor.