Trait unsafe_io::AsUnsafeReadWriteHandle

pub trait AsUnsafeReadWriteHandle {
    fn as_unsafe_read_handle(&self) -> UnsafeHandle;

    fn as_unsafe_write_handle(&self) -> UnsafeHandle;
}

Like AsUnsafeHandle, but for types which may have one or two handles, for reading and writing.

For types that only have one, both functions return the same value.

Required methods

fn as_unsafe_read_handle(&self) -> UnsafeHandle

Extracts the unsafe handle for reading.

fn as_unsafe_write_handle(&self) -> UnsafeHandle

Extracts the unsafe handle for writing.

Loading content...